+++ src/num/mod.rs
@@ -1,6 +1,6 @@
 // devela::num
 //
-//! Numerical types and operations, algebra, geometry, <small>extends
+//! Numerical types and operations, algebra, <small>extends
 //! `std::{`[`cmp`], [`num`]`}`.</small>
 //!
 //! [`cmp`]: std::cmp
@@ -22,11 +22,13 @@ mod no;
 mod primitive;
 mod sign;
 mod r#trait;
-pub use {alias::*, cmp::*, error::*, float::*, no::*, primitive::*, r#trait::*, sign::*};
+mod unit;
+pub use {alias::*, cmp::*, error::*, float::*, no::*, primitive::*, r#trait::*, sign::*, unit::*};
 
+pub mod algebra;
 pub mod niche;
 #[doc(no_inline)]
-pub use niche::all::*;
+pub use {algebra::all::*, niche::all::*};
 
 #[cfg(_some_int)]
 #[cfg_attr(feature = "nightly_doc", doc(cfg(_some_int)))]
@@ -36,12 +38,6 @@ mod frac;
 mod int;
 #[cfg(_some_int)]
 pub use {frac::*, int::*};
-
-#[cfg(feature = "num_geom")]
-#[cfg_attr(feature = "nightly_doc", doc(cfg(feature = "num_geom")))]
-pub mod geom;
-#[cfg(feature = "num_geom")]
-pub use geom::*;
 
 #[cfg(feature = "num_rand")]
 #[cfg_attr(feature = "nightly_doc", doc(cfg(feature = "num_rand")))]
@@ -54,14 +50,9 @@ pub(crate) mod all {
     #[doc(inline)]
     #[allow(unused_imports)]
     pub use super::{
-        alias::*, cmp::all::*, error::*, float::*, niche::all::*, no::*, primitive::*, r#trait::*,
-        sign::*,
+        algebra::all::*, alias::*, cmp::all::*, error::*, float::*, niche::all::*, no::*,
+        primitive::*, r#trait::*, sign::*, unit::*,
     };
-
-    #[doc(inline)]
-    #[cfg(feature = "num_geom")]
-    #[allow(unused_imports)]
-    pub use super::geom::all::*;
 
     #[doc(inline)]
     #[cfg(_some_int)]